requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
psycopg2-binary
python-dotenv
opencv-python-headless
numpy
imutils